change submit route to server action
